requirements.txt
# Core dependencies
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
openai>=1.0.0
chromadb>=0.4.0

# PDF processing
pymupdf>=1.23.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
